@@ -1,67 +1,66 @@
 {
   "name": "angular-web",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "private": false,
   "scripts": {
-    "lint": "tslint ./src/**/*.ts -t verbose",
-    "start": "node server.js",
-    "start:dev": "ng serve --port 8080",
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build --configuration=production",
     "test": "ng test",
     "test-headless": "ng test --watch=false --browsers=ChromeHeadless",
-    "build": "ng build --prod",
-    "e2e": "ng e2e --port 0",
-    "ng": "ng"
+    "lint": "ng lint",
+    "lint:fix": "ng lint --fix",
+    "prettier": "prettier --write \"src/**/*.{js,ts,html,scss,json}",
+    "e2e": "ng e2e"
   },
-  "keywords": [],
-  "author": "Peter Patocka",
-  "repository": "https://github.com/peter-patocka/angular-web",
-  "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
-    "@angular/upgrade": "^8.0.0",
-    "@fortawesome/fontawesome-free": "^5.11.2",
-    "angular-in-memory-web-api": "^0.9.0",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
-    "express": "^4.17.1",
-    "jquery": "^3.5.0",
-    "path": "^0.12.7",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.5.1",
-    "typescript": "~3.4.3",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "^17.2.3",
+    "@angular/common": "^17.2.3",
+    "@angular/compiler": "^17.2.3",
+    "@angular/core": "^17.2.3",
+    "@angular/forms": "^17.2.3",
+    "@angular/platform-browser": "^17.2.3",
+    "@angular/platform-browser-dynamic": "^17.2.3",
+    "@angular/router": "^17.2.3",
+    "angular-in-memory-web-api": "^0.17.0",
+    "core-js": "^3.36.0",
+    "font-awesome": "^4.7.0",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.6.2",
+    "wowjs": "^1.1.3",
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.800.0",
-    "@angular/cli": "^8.3.19",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/platform-server": "^8.2.14",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "^2.0.4",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.2.0",
-    "enhanced-resolve": "^3.3.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "lodash": "^4.16.2",
-    "protractor": "^5.4.2",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0"
-  },
-  "engines": {
-    "node": "12.13.1",
-    "npm": "6.12.1"
+    "@angular-devkit/build-angular": "^17.2.2",
+    "@angular-eslint/builder": "17.2.1",
+    "@angular-eslint/eslint-plugin": "17.2.1",
+    "@angular-eslint/eslint-plugin-template": "17.2.1",
+    "@angular-eslint/schematics": "17.2.1",
+    "@angular-eslint/template-parser": "17.2.1",
+    "@angular/cli": "^17.2.2",
+    "@angular/compiler-cli": "^17.2.3",
+    "@angular/language-service": "^17.2.3",
+    "@types/flickity": "^2.2.11",
+    "@types/jasmine": "~5.1.4",
+    "@types/jasminewd2": "~2.0.13",
+    "@types/node": "^20.11.24",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
+    "codelyzer": "^6.0.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "jasmine-core": "~5.1.2",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.3",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "prettier": "^3.2.5",
+    "prettier-eslint": "^16.3.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~10.9.2",
+    "typescript": "5.2.2"
   }
 }
