@@ -1,50 +1,44 @@
-import { HttpClientModule } from '@angular/common/http';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
 import { HttpClientInMemoryWebApiModule } from 'angular-in-memory-web-api';
-import { AppRoutingModule } from './app-routing.module';
+import { ReactiveFormsModule } from '@angular/forms';
+
+// Components
 import { AppComponent } from './app.component';
-import { AuthModule } from './auth/auth.module';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { HeroDetailComponent } from './hero-detail/hero-detail.component';
-import { HeroSearchComponent } from './hero-search/hero-search.component';
-import { HeroesComponent } from './heroes/heroes.component';
-import { NavigationComponent } from './heroes/navigation/navigation.component';
-import { HomeComponent } from './home/home.component';
-import { InMemoryDataService } from './in-memory-data.service';
-import { MessagesComponent } from './messages/messages.component';
-import { TestComponent } from './test/test.component';
-import { TopComponent } from './top/top.component';
+import { FooterComponent } from './footer/footer.component';
+import { NavmenuComponent } from './navmenu/navmenu.component';
+import { NavigationComponent } from './navigation/navigation.component';
+import { SocialComponent } from './social/social.component';
+
+// Modules
+import { AppRoutingModule } from './app-routing.module';
+
+// Services
+import { ConfigService } from './shared/services/config.service';
+import { InMemoryDataService } from './shared/services/in-memory-data.service';
 
 @NgModule({
-  imports: [
-    BrowserModule,
-    FormsModule,
-    AppRoutingModule,
-    HttpClientModule,
-    AuthModule,
+	declarations: [AppComponent],
+	imports: [
+		AppRoutingModule,
+		BrowserModule,
+		HttpClientModule,
+		ReactiveFormsModule,
 
-    // The HttpClientInMemoryWebApiModule module intercepts HTTP requests
-    // and returns simulated server responses.
-    // Remove it when a real server is ready to receive requests.
-    HttpClientInMemoryWebApiModule.forRoot(
-      InMemoryDataService, { dataEncapsulation: false }
-    )
-  ],
-  declarations: [
-    AppComponent,
-    DashboardComponent,
-    HeroesComponent,
-    HeroDetailComponent,
-    MessagesComponent,
-    HeroSearchComponent,
-    NavigationComponent,
-    TestComponent,
-    TopComponent,
-    HomeComponent
-  ],
-  bootstrap: [AppComponent]
+		// The HttpClientInMemoryWebApiModule module intercepts HTTP requests
+		// and returns simulated server responses.
+		// Remove it when a real server is ready to receive requests.
+		HttpClientInMemoryWebApiModule.forRoot(InMemoryDataService, {
+			dataEncapsulation: false,
+			passThruUnknownUrl: true,
+		}),
+		FooterComponent,
+		NavigationComponent,
+		NavmenuComponent,
+		SocialComponent,
+	],
+	providers: [ConfigService],
+	bootstrap: [AppComponent],
 })
-
-export class AppModule { }
+export class AppModule {}
