@@ -1,4 +1,3 @@
 export const environment = {
-  production: true,
-  baseHref: '/angular-web/'
+	production: true,
 };
